services/crawler/requirements.txt
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.5.7
charset-normalizer==3.2.0
greenlet==2.0.2
idna==3.4
numpy==1.24.4
Pillow==10.0.0
psycopg2==2.9.6
requests==2.31.0
soupsieve==2.4.1
SQLAlchemy==2.0.18
typing-extensions==4.7.1
urllib3==2.0.3
